+++ app/views/incomeSources/manage/ManageObligations.scala.html
@@ -18,7 +18,7 @@
 @import views.html.layouts.unifiedLayout
 @import views.html.components._
 @import uk.gov.hmrc.govukfrontend.views.html.components._
-@import models.TaxYearId
+@import models.core.TaxYearId
 @import models.incomeSourceDetails.viewmodels.ObligationsViewModel
 
 @this(
